LayoutTests/platform/blackberry/editing/text-iterator/findString-markers.html

<!DOCTYPE html>
<html>
<head>
<style>
.plaintext {
    border: 2px solid red;
    padding: 12px;
    font-size: 24px;
}
</style>
<script src="../../../../fast/js/resources/js-test-pre.js"></script>
</head>
<body>
<pre id="description"></pre>
<pre id="console"></pre>
<div id="container">
</div>

<script>

if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
}

function createPlainTextElement(id, text, parent) {
    var e = document.createElement('div');
    e.className = 'plaintext';
    e.id = id;
    e.innerText = text;

    parent.appendChild(e);
    return e;
}

function occurrences(string, substring, caseSensitive) {
    var n=0;
    var pos=0;

    var haystack = caseSensitive ? string : string.toLowerCase();
    var needle = caseSensitive ? substring : substring.toLowerCase();

    while (true) {
        pos = haystack.indexOf(needle,pos);
        if (pos!=-1) {
            ++n;
            pos+=needle.length;
        } else
            break;
    }
    return n;
}

function checkTextMatchMarker(index, findOptions) {
    if (index > 12) {
        if (findOptions.length == 0) {
            index = 0;
            findOptions.push("CaseInsensitive");
        } else {
            testRunner.notifyDone();
            return;
        }
    }
    var str = 'ABCdefTUVxyz.';
    var char = str.charAt(index);
    var caseSensitive = findOptions.indexOf("CaseInsensitive") == -1;
    testRunner.findString(char, findOptions);
    // Start the function later to allow marking process to finish.
    setTimeout(function() {
        var count = occurrences(text, char, caseSensitive);

        // The number of markers should match to the occurrences of the charactor.
        shouldBe('internals.markerCountForNode(e.firstChild, "textmatch")', count.toString());
        var searchStartIndex = 0;
        // Get the char from text and compare with the marker range content.
        for (var i = 0; i < count ; ++i) {
            var range = internals.markerRangeForNode(e.firstChild, "textmatch", i);
            var pos = caseSensitive ? text.indexOf(char, searchStartIndex)
                                    : text.toLowerCase().indexOf(char.toLowerCase(), searchStartIndex);
            searchStartIndex = pos + 1;
            var expectStr = text.charAt(pos);
            shouldBe('"' + range + '"', '"' + expectStr + '"');
        }
        checkTextMatchMarker(index + 1, findOptions);
    }, 850);
}

var container = document.getElementById('container');

var id = 'id1'
var text = 'The quick brown fox jumps over the lazy dog.';
var e = createPlainTextElement(id, text, container);

if (window.internals && window.testRunner)
    checkTextMatchMarker(0, []);

var successfullyParsed = true;
</script>
<script src="../../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
